refactor away from macros
