+++ src/backend/cargo.rs
@@ -26,7 +26,7 @@ impl Backend for Cargo {
     type QueryInfo = CargoInstallOptions;
     type Modification = ();
 
-    fn query_installed_packages(_: &Config) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
+    fn query_installed_packages(&self,_: &Config) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
         if !command_found("cargo") {
             return Ok(BTreeMap::new());
         }
@@ -48,6 +48,7 @@ impl Backend for Cargo {
     }
 
     fn install_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::InstallOptions>,
         _: bool,
         _: &Config,
@@ -81,6 +82,7 @@ impl Backend for Cargo {
     }
 
     fn modify_packages(
+        &self,
         _: &BTreeMap<Self::PackageId, Self::Modification>,
         _: &Config,
     ) -> Result<()> {
@@ -88,6 +90,7 @@ impl Backend for Cargo {
     }
 
     fn remove_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::RemoveOptions>,
         _: bool,
         _: &Config,
