+++ src/backend/dnf.rs
@@ -26,7 +26,7 @@ impl Backend for Dnf {
     type QueryInfo = DnfQueryInfo;
     type Modification = ();
 
-    fn query_installed_packages(_: &Config) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
+    fn query_installed_packages(&self,_: &Config) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
         if !command_found("dnf") {
             return Ok(BTreeMap::new());
         }
@@ -56,6 +56,7 @@ impl Backend for Dnf {
     }
 
     fn install_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::InstallOptions>,
         no_confirm: bool,
         _: &Config,
@@ -78,6 +79,7 @@ impl Backend for Dnf {
     }
 
     fn modify_packages(
+        &self,
         _: &BTreeMap<Self::PackageId, Self::Modification>,
         _: &Config,
     ) -> Result<()> {
@@ -85,6 +87,7 @@ impl Backend for Dnf {
     }
 
     fn remove_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::RemoveOptions>,
         no_confirm: bool,
         _: &Config,
