+++ src/backend/flatpak.rs
@@ -21,7 +21,7 @@ impl Backend for Flatpak {
     type QueryInfo = FlatpakQueryInfo;
     type Modification = ();
 
-    fn query_installed_packages(_: &Config) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
+    fn query_installed_packages(&self, _: &Config) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
         if !command_found("flatpak") {
             return Ok(BTreeMap::new());
         }
@@ -111,6 +111,7 @@ impl Backend for Flatpak {
     }
 
     fn install_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::InstallOptions>,
         no_confirm: bool,
         config: &Config,
@@ -132,6 +133,7 @@ impl Backend for Flatpak {
     }
 
     fn modify_packages(
+        &self,
         _: &BTreeMap<Self::PackageId, Self::Modification>,
         _: &Config,
     ) -> Result<()> {
@@ -139,6 +141,7 @@ impl Backend for Flatpak {
     }
 
     fn remove_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::RemoveOptions>,
         no_confirm: bool,
         config: &Config,
