+++ src/backend/pacman.rs
@@ -16,11 +16,13 @@ impl Backend for Pacman {
     type Modification = ArchModification;
 
     fn query_installed_packages(
+        &self,
         config: &Config,
     ) -> Result<BTreeMap<Self::PackageId, Self::QueryInfo>> {
         Self::PACMAN.query_installed_packages(config)
     }
     fn install_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::InstallOptions>,
         no_confirm: bool,
         config: &Config,
@@ -28,12 +30,14 @@ impl Backend for Pacman {
         Self::PACMAN.install_packages(packages, no_confirm, config)
     }
     fn modify_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::Modification>,
         config: &Config,
     ) -> Result<()> {
         Self::PACMAN.modify_packages(packages, config)
     }
     fn remove_packages(
+        &self,
         packages: &BTreeMap<Self::PackageId, Self::RemoveOptions>,
         no_confirm: bool,
         config: &Config,
